Fix process.Finalize under unix when SysProcAttr is nil

+++ modules/process/unix.go
@@ -1,4 +1,4 @@
-// +build aix darwin dragonfly freebsd linux netbsd openbsd solaris
+//go:build aix || darwin || dragonfly || freebsd || linux || netbsd || openbsd || solaris
 
 package process
 
@@ -11,8 +11,10 @@ func Finalize(cmd *exec.Cmd) error {
 	if cmd == nil || cmd.Process == nil || cmd.Process.Pid <= 0 {
 		return nil
 	}
-	if cmd.SysProcAttr.Setpgid {
-		syscall.Kill(-cmd.Process.Pid, syscall.SIGTERM)
+	if cmd.SysProcAttr != nil {
+		if cmd.SysProcAttr.Setpgid {
+			syscall.Kill(-cmd.Process.Pid, syscall.SIGTERM)
+		}
 	}
 	syscall.Kill(cmd.Process.Pid, syscall.SIGTERM)
 	return cmd.Wait()
